Tratamento dos erros personalizados

+++ api/src/model/usuario.go
@@ -4,7 +4,7 @@ import "time"
 
 // Usuario - modelo de um usuário
 type Usuario struct {
-	ID       uint      `json:"id,omitempty"`
+	ID       uint64    `json:"id,omitempty"`
 	Nome     string    `json:"nome,omitempty"`
 	Nick     string    `json:"nick,omitempty"`
 	Email    string    `json:"email,omitempty"`
